#354 Make string definitions self resolvable
