+++ src/DI/Definition/StringDefinition.php
@@ -2,7 +2,10 @@
 
 namespace DI\Definition;
 
+use DI\DependencyException;
 use DI\Scope;
+use Interop\Container\ContainerInterface;
+use Interop\Container\Exception\NotFoundException;
 
 /**
  * Definition of a string composed of other strings.
@@ -10,7 +13,7 @@ use DI\Scope;
  * @since 5.0
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
-class StringDefinition implements Definition
+class StringDefinition implements Definition, SelfResolvingDefinition
 {
     /**
      * Entry name.
@@ -55,5 +58,33 @@ class StringDefinition implements Definition
     public function getExpression()
     {
         return $this->expression;
+    }
+
+    public function resolve(ContainerInterface $container)
+    {
+        $expression = $this->expression;
+
+        $result = preg_replace_callback('#\{([^\{\}]+)\}#', function (array $matches) use ($container) {
+            try {
+                return $container->get($matches[1]);
+            } catch (NotFoundException $e) {
+                throw new DependencyException(sprintf(
+                    "Error while parsing string expression for entry '%s': %s",
+                    $this->getName(),
+                    $e->getMessage()
+                ), 0, $e);
+            }
+        }, $expression);
+
+        if ($result === null) {
+            throw new \RuntimeException(sprintf('An unknown error occurred while parsing the string definition: \'%s\'', $expression));
+        }
+
+        return $result;
+    }
+
+    public function isResolvable(ContainerInterface $container)
+    {
+        return true;
     }
 }
